Move some integration tests
